Add cast

+++ mx_thread.c
@@ -257,5 +257,5 @@ void mx_semaphore_signal(mx_semaphore *s, uint32_t num)
 
 int32_t mx_semaphore_get_count(const mx_semaphore *s)
 {
-	return mxa_load32(&s->count);
+	return (int32_t)mxa_load32(&s->count);
 }
